+++ webpack.config.js
@@ -50,10 +50,8 @@ module.exports = function(opts) {
     }
   };
 
-  // Enable sourcemaps as long as we're not in production mode
-  if (!production) {
-    config.devtool = 'eval-source-maps';
-  }
+  // Enable sourcemaps, different kinds depending on environment
+  config.devtool = production ? 'source-map' : 'eval-source-maps';
 
   /***********
    * LOADERS *